+++ src/renderer/src/App.tsx
@@ -20,6 +20,7 @@ import { EvidencePage } from './pages/Evidence'
 import { Enumerate } from './pages/Enumerate'
 import { Timeline } from './pages/Timeline'
 import { ExamPrep } from './pages/ExamPrep'
+import { MapView } from './pages/MapView'
 import { Vpn } from './pages/Vpn'
 import { WhatsNew } from './pages/WhatsNew'
 import { Settings } from './pages/Settings'
@@ -54,6 +55,7 @@ export default function App(): JSX.Element {
               <Route path="/enumerate" element={<Enumerate />} />
               <Route path="/timeline" element={<Timeline />} />
               <Route path="/exam-prep" element={<ExamPrep />} />
+              <Route path="/map" element={<MapView />} />
               <Route path="/graph" element={<Graph />} />
               <Route path="/vpn" element={<Vpn />} />
               <Route path="/whats-new" element={<WhatsNew />} />